chore: Update sidebar, add icons refactors
