+++ src/components/FooterComponent.jsx
@@ -2,11 +2,13 @@ import React from "react";
 
 const FooterComponent = () => {
   return (
-    <footer className="bg-white text-center py-5 border-t-2 text-neutral-800">
-      <span className="font-bold text-black">
-        Record<span className="text-green-600">Finance</span>
-      </span>{" "}
-      Copyright © {new Date().getFullYear()}
+    <footer>
+      <p className="flex justify-center py-3 italic opacity-25 gap-1 text-xs">
+        <span className="font-bold text-black">
+          Record<span className="text-green-600">Finance</span>
+        </span>{" "}
+        Copyright © {new Date().getFullYear()}
+      </p>
     </footer>
   );
 };
