+++ src/components/InputComponent.jsx
@@ -7,17 +7,23 @@ const InputComponent = ({
   placeholder,
   onChange,
   props,
+  icon,
 }) => {
   return (
-    <input
-      id={id}
-      type={type}
-      name={name}
-      className="border w-full p-2 mt-2 bg-gray-50 rounded-xl"
-      placeholder={placeholder}
-      onChange={onChange}
-      {...props}
-    />
+    <div className="relative">
+      <span className="absolute top-4 left-2">{icon}</span>
+      <input
+        id={id}
+        type={type}
+        name={name}
+        className={`${
+          icon ? "pl-10" : ""
+        } border w-full p-2 mt-2 bg-gray-50 rounded-xl`}
+        placeholder={placeholder}
+        onChange={onChange}
+        {...props}
+      />
+    </div>
   );
 };
