+++ src/components/index.js
@@ -1,8 +1,14 @@
-export { default as Input } from "./InputComponent";
-export { default as Label } from "./LabelComponent";
-export { default as Button } from "./ButtonComponent";
+// Page
 export { default as Header } from "./HeaderComponent";
 export { default as Footer } from "./FooterComponent";
-export { default as Sidebar } from "./SidebarComponent";
-export { default as ItemMenu } from "./ItemMenuComponent";
+
+// Component's System
+export { default as Input } from "./InputComponent";
+export { default as Label } from "./LabelComponent";
 export { default as Loading } from "./LoadingComponent";
+export { default as Button } from "./ButtonComponent";
+
+// sidebar
+export { default as Sidebar } from "./sidebar/SidebarComponent";
+export { default as ItemMenu } from "./sidebar/ItemMenuComponent";
+export { default as SectionMenu } from "./sidebar/SectionMenuComponent";
