+++ src/helpers/icons.jsx
@@ -1,51 +1,165 @@
 export const dashboard = (
   <svg
-    className="w-6 h-6 fill-current text-white"
-    enableBackground="new 0 0 512 512"
-    viewBox="0 0 512 512"
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
   >
-    <path d="m256.01 42.671h-7e-3 -2e-3 -3e-3c-70.671 3e-3 -134.65 28.637-180.97 74.935-0.016 0.016-0.034 0.029-0.05 0.045-0.017 0.017-0.03 0.035-0.047 0.052-46.245 46.273-74.861 110.16-74.922 180.74 0 0.076-0.011 0.15-0.011 0.226 0 0.073 0.01 0.143 0.011 0.215 0.05 60.201 20.962 117.24 58.515 162.7 4.053 4.907 10.084 7.748 16.448 7.748h362.05c6.364 0 12.395-2.841 16.448-7.748 37.607-45.53 58.539-102.65 58.539-162.92 0-141.38-114.62-256-256-256zm170.67 384h-341.36c-23.281-30.977-37.712-67.661-41.583-106.67h62.934c11.782 0 21.333-9.551 21.333-21.333s-9.551-21.333-21.333-21.333h-62.933c4.259-42.905 21.23-82.066 47.091-113.67l14.32 14.32c8.331 8.331 21.839 8.331 30.17 0s8.331-21.839 0-30.17l-14.321-14.321c31.605-25.864 70.765-42.837 113.67-47.098v62.941c0 11.782 9.551 21.333 21.333 21.333s21.333-9.551 21.333-21.333v-62.943c42.906 4.259 82.068 21.232 113.68 47.096l-14.325 14.325c-8.331 8.331-8.331 21.839 0 30.17s21.839 8.331 30.17 0l14.326-14.326c25.867 31.607 42.842 70.771 47.103 113.68h-62.95c-11.782 0-21.333 9.551-21.333 21.333s9.551 21.333 21.333 21.333h62.95c-3.873 38.997-18.312 75.682-41.602 106.67z" />
-    <polygon points="319.99 220.66 199.33 281.01 259.66 341.34" />
-    <path d="m320 362.67h-128c-11.782 0-21.333 9.551-21.333 21.333s9.551 21.333 21.333 21.333h128c11.782 0 21.333-9.551 21.333-21.333s-9.551-21.333-21.333-21.333z" />
+    <path stroke="currentColor" d="M3,16V15a9,9,0,0,1,9-9h0a9,9,0,0,1,9,9v1" />
+    <circle stroke="currentColor" cx="12" cy="16" r="2" />
+    <line stroke="currentColor" x1="13.41" y1="14.59" x2="16.5" y2="11.5" />
   </svg>
 );
 
 export const wallet = (
   <svg
-    className="w-6 h-6 fill-current text-white"
-    enableBackground="new 0 0 512 512"
-    viewBox="0 0 512 512"
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
   >
-    <path d="m59.733 477.85h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m162.13 477.85h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m162.13 170.65h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m213.33 477.85h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m110.93 170.65h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m59.733 170.65h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m486.4 375.45c-4.71 0-8.533 3.823-8.533 8.533v85.333c0 14.114-11.486 25.6-25.6 25.6h-409.6c-14.114 0-25.6-11.486-25.6-25.6v-341.33c0-15.522 14.43-26.522 28.049-30.225l256.74-77.056c5.163-2.082 21.99-6.622 31.275-0.341 5.444 3.678 8.201 11.179 8.201 22.289 0 4.71 3.823 8.533 8.533 8.533s8.533-3.823 8.533-8.533c0-17.126-5.291-29.389-15.718-36.429-18.355-12.399-44.305-2.517-46.464-1.613l-255.79 76.74c-20.13 5.47-40.431 22.016-40.431 46.635v341.33c0 23.526 19.14 42.667 42.667 42.667h409.6c23.526 0 42.667-19.14 42.667-42.667v-85.333c-1e-3 -4.71-3.824-8.533-8.534-8.533z" />
-    <path d="m213.33 170.65h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m110.93 477.85h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m494.93 273.05v-110.93c0-23.526-19.14-42.667-42.667-42.667h-319.2l269.76-51.379c0.085-0.026 9.779-1.911 16.393 3.49 4.941 4.036 7.441 11.537 7.441 22.289 0 4.71 3.823 8.533 8.533 8.533s8.533-3.823 8.533-8.533c0-16.239-4.676-28.237-13.892-35.661-13.039-10.487-29.892-6.963-30.37-6.852l-358.4 68.267c-4.326 0.819-7.313 4.804-6.903 9.182 0.418 4.378 4.096 7.731 8.499 7.731h409.6c14.114 0 25.6 11.486 25.6 25.6v110.93h-170.67c-9.412 0-17.067 7.654-17.067 17.067v51.2c0 9.412 7.654 17.067 17.067 17.067h187.73c9.412 0 17.067-7.654 17.067-17.067v-51.2c0-9.412-7.654-17.066-17.067-17.066zm-187.73 68.266v-51.2h187.73l9e-3 51.2h-187.74z" />
-    <path d="m384 170.65c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533h17.067z" />
-    <path d="m332.8 170.65c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533h17.067z" />
-    <path d="m435.2 170.65c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533h17.067z" />
-    <path d="m333.06 307.18h-0.085c-4.71 0-8.491 3.823-8.491 8.533s3.866 8.533 8.576 8.533c4.719 0 8.533-3.823 8.533-8.533s-3.814-8.533-8.533-8.533z" />
-    <path d="m332.8 477.85c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533h17.067z" />
-    <path d="m435.2 477.85c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533h17.067z" />
-    <path d="m264.53 170.65h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m264.53 477.85h17.067c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533z" />
-    <path d="m384 477.85c4.71 0 8.533-3.823 8.533-8.533s-3.823-8.533-8.533-8.533h-17.067c-4.71 0-8.533 3.823-8.533 8.533s3.823 8.533 8.533 8.533h17.067z" />
+    <rect stroke="currentColor" x="2" y="6" width="20" height="16" rx="2" />
+    <path stroke="currentColor" d="M18,6V4.66a2,2,0,0,0-2.55-1.93L2.88,6.34" />
+    <path
+      stroke="currentColor"
+      d="M22,11.64H18.18A2.18,2.18,0,0,0,16,13.82h0A2.19,2.19,0,0,0,18.18,16H22"
+    />
   </svg>
 );
 
 export const purchase = (
   <svg
-    className="w-6 h-6 fill-current text-white"
-    enableBackground="new 0 0 512 512"
-    viewBox="0 0 512 512"
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
   >
-    <path d="m134.33 212.11c9.592 7.664 21.632 11.888 33.904 11.888h15.768v8c0 13.232 10.768 24 24 24h104c13.232 0 24-10.768 24-24v-40h8c13.232 0 24-10.768 24-24 0-6.168-2.408-11.744-6.24-16 3.84-4.256 6.24-9.832 6.24-16s-2.408-11.744-6.24-16c3.84-4.256 6.24-9.832 6.24-16s-2.408-11.744-6.24-16c3.84-4.256 6.24-9.832 6.24-16 0-13.232-10.768-24-24-24h-8v-24c0-13.232-10.768-24-24-24h-104c-13.232 0-24 10.768-24 24v16h-13.072c-8.384 0-16.776 1.976-24.28 5.728l-36.536 18.272h-110.11v16h113.89l39.912-19.96c5.296-2.64 11.216-4.04 17.128-4.04h13.072v118.78l36.584 6.104c6.616 1.096 11.416 6.76 11.416 13.464 0 7.528-6.128 13.656-13.656 13.656h-50.112c-8.656 0-17.152-2.976-23.912-8.392l-29.512-23.608h-114.81v16h109.19l25.136 20.112zm201.67-148.11h8c4.416 0 8 3.584 8 8s-3.584 8-8 8h-8v-16zm0 32h8c4.416 0 8 3.584 8 8s-3.584 8-8 8h-8v-16zm0 32h8c4.416 0 8 3.584 8 8s-3.584 8-8 8h-8v-16zm0 32h8c4.416 0 8 3.584 8 8s-3.584 8-8 8h-8v-16zm-32-144h8c4.416 0 8 3.584 8 8v208c0 4.416-3.584 8-8 8h-8v-224zm-104 208h18.344c16.352 0 29.656-13.304 29.656-29.656 0-14.552-10.424-26.848-24.784-29.248l-23.216-3.872v-137.22c0-4.416 3.584-8 8-8h80v224h-80c-4.416 0-8-3.584-8-8v-8z" />
-    <path d="m272 80c0-4.072-0.832-8.152-2.424-12 1.592-3.848 2.424-7.928 2.424-12 0-17.648-14.352-32-32-32s-32 14.352-32 32c0 4.072 0.824 8.152 2.424 12-1.6 3.848-2.424 7.928-2.424 12 0 17.648 14.352 32 32 32s32-14.352 32-32zm-18.24-7.968c1.488 2.544 2.24 5.224 2.24 7.968 0 8.816-7.184 16-16 16-8.824 0-16-7.184-16-16 0-2.744 0.752-5.424 2.24-7.968l2.344-4.032-2.344-4.032c-1.488-2.544-2.24-5.224-2.24-7.968 0-8.816 7.176-16 16-16 8.816 0 16 7.184 16 16 0 2.744-0.752 5.424-2.24 7.968l-2.344 4.032 2.344 4.032z" />
-    <path d="m278.19 312h49.568c8.656 0 17.152 2.976 23.912 8.392l29.52 23.608h114.81v-16h-109.19l-25.136-20.112c-9.592-7.664-21.632-11.888-33.904-11.888h-7.768v-24h-88v8c0 17.648-14.352 32-32 32s-32-14.352-32-32v-8h-88v80h32v144h176v-16h37.072c8.384 0 16.776-1.976 24.28-5.728l36.536-18.272h110.11v-16h-113.89l-39.912 19.96c-5.296 2.64-11.216 4.04-17.128 4.04h-37.072v-104.7l38.264 8.504 3.472-15.624-54.624-12.136c-6.544-1.448-11.112-7.144-11.112-13.856 0-7.824 6.368-14.192 14.192-14.192zm-6.192 168h-144v-80h144v80zm0-112v16h-144v-16h144zm-9.472-16h-134.53v-16h-32v-48h56.664c3.816 22.672 23.592 40 47.336 40s43.512-17.328 47.336-40h56.664v8h-25.808c-16.648 0-30.192 13.544-30.192 30.192 0 10.88 5.672 20.48 14.528 25.808z" />
+    <polyline
+      stroke="currentColor"
+      points="1.68 2.77 5.18 2.77 8.34 16.48 21 16.48 22.32 6.72 6.09 6.72"
+    />
+    <rect
+      stroke="currentColor"
+      x="9.3"
+      y="18.62"
+      width="2.61"
+      height="2.61"
+      rx="1.3"
+    />
+    <rect
+      stroke="currentColor"
+      x="9.3"
+      y="18.62"
+      width="2.61"
+      height="2.61"
+      rx="1.3"
+    />
+    <rect
+      stroke="currentColor"
+      x="16.16"
+      y="18.62"
+      width="2.61"
+      height="2.61"
+      rx="1.3"
+    />
+  </svg>
+);
+
+export const chevronRight = (
+  <svg
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+  >
+    <path
+      stroke="currentColor"
+      d="M9.8,16.78l4.11-4.11a1,1,0,0,0,0-1.41l-4-4"
+    />
+  </svg>
+);
+
+export const logOut = (
+  <svg
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+  >
+    <path
+      stroke="currentColor"
+      d="M15.71,15v4a2,2,0,0,1-2,2h-6a2,2,0,0,1-2-2V5a2,2,0,0,1,2-2h6a2,2,0,0,1,2,2V9"
+    />
+    <line stroke="currentColor" x1="12.5" y1="11.95" x2="22.5" y2="11.95" />
+    <path
+      stroke="currentColor"
+      d="M18.82,8l3.44,3.44a.83.83,0,0,1,0,1.18L18.88,16"
+    />
+  </svg>
+);
+
+export const trash = (
+  <svg
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+  >
+    <line stroke="currentColor" x1="15" y1="17.2" x2="9" y2="11.2" />
+    <line stroke="currentColor" x1="15" y1="11.2" x2="9" y2="17.2" />
+    <rect stroke="currentColor" x="3" y="1.81" width="18" height="4" rx="1" />
+    <path
+      stroke="currentColor"
+      d="M19,5.81V20.19a2,2,0,0,1-2,2H7a2,2,0,0,1-2-2V5.81"
+    />
+  </svg>
+);
+
+export const edit = (
+  <svg
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+  >
+    <path
+      stroke="currentColor"
+      d="M16.74,4.55a1.63,1.63,0,0,0-2.29.07l-10,10v0l-.31,4.86a.33.33,0,0,0,.35.35l4.86-.31h0l10-10a1.63,1.63,0,0,0,.07-2.29Z"
+    />
+  </svg>
+);
+
+export const envelop = (
+  <svg
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+  >
+    <rect stroke="currentColor" x="2" y="5" width="20" height="14" rx="2" />
+    <path stroke="currentColor" d="M2.58,5.59l8.23,8.22a2,2,0,0,0,2.83,0l8-8" />
+  </svg>
+);
+
+export const key = (
+  <svg
+    width="24px"
+    height="24px"
+    viewBox="0 0 24 24"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+  >
+    <circle stroke="currentColor" cx="6" cy="12" r="4" />
+    <line stroke="currentColor" x1="10" y1="12" x2="22" y2="12" />
+    <line stroke="currentColor" x1="17" y1="12" x2="17" y2="15" />
+    <line stroke="currentColor" x1="21" y1="12" x2="21" y2="14" />
   </svg>
 );
